tabfpn/requirements.txt
tabpfn-client
scikit-learn>=1.4
numpy>=1.26
python-dotenv
